order-bot/requirements.txt
faust-streaming==0.11.3
sqlalchemy==2.0.36
discord_webhook==1.3.1
asyncpg==0.30.0
greenlet==3.1.1